+++ src/redux/actions/auth.js
@@ -1,0 +1,6 @@
+import { LOG_IN } from './actionTypes'
+
+export const logIn = auth => ({
+  type: LOG_IN,
+  payload: auth
+})
