+++ src/redux/actions/movies.js
@@ -1,0 +1,6 @@
+import { ADD_MOVIE } from './actionTypes'
+
+export const addMovie = movie => ({
+  type: ADD_MOVIE,
+  payload: movie
+})
