+++ src/redux/actions/shows.js
@@ -1,0 +1,6 @@
+import { ADD_SHOW } from './actionTypes'
+
+export const addShow = show => ({
+  type: ADD_SHOW,
+  payload: show
+})
