+++ src/redux/reducers/movies.js
@@ -1,9 +1,9 @@
-import { ADD_MOVIE } from '../actionTypes'
+import { ADD_MOVIE } from '../actions/actionTypes'
 
 export default (state = {}, action) => {
   switch (action.type) {
   case ADD_MOVIE:
-    return state[action.payload.id] = action.payload.movie
+    return state[action.payload.id] = action.payload
   default:
     return state
   }
