+++ src/redux/selectors/auth.js
@@ -1,1 +1,3 @@
+// import { createSelector } from 'reselect'
+
 export const getUserState = state => state.auth.user
